+++ src/pages/Tag/index.tsx
@@ -1,12 +1,15 @@
 import { useState, useEffect } from 'react';
 import { Table, Button, Form, Input, Popconfirm, message, Card } from 'antd';
-import { getTagListAPI, addTagDataAPI, editTagDataAPI, delTagDataAPI } from '@/api/Tag';
+import { getTagListAPI, addTagDataAPI, editTagDataAPI, delTagDataAPI, getTagDataAPI } from '@/api/Tag';
 import { Tag } from '@/types/app/tag';
 import Title from '@/components/Title';
 import { ColumnsType } from 'antd/es/table';
 
 const TagPage = () => {
     const [loading, setLoading] = useState<boolean>(false);
+    const [btnLoading, setBtnLoading] = useState(false)
+
+    const [form] = Form.useForm();
 
     const [tag, setTag] = useState<Tag>({} as Tag);
     const [list, setList] = useState<Tag[]>([]);
@@ -28,49 +31,76 @@ const TagPage = () => {
     ];
 
     const getTagList = async () => {
-        setLoading(true);
-        const { data } = await getTagListAPI();
-        setList(data as Tag[]);
+        try {
+            const { data } = await getTagListAPI();
+            setList(data as Tag[]);
+        } catch (error) {
+            setLoading(false);
+        }
+
         setLoading(false);
     };
 
     useEffect(() => {
+        setLoading(true);
         getTagList();
     }, []);
 
-    const [form] = Form.useForm();
-    const editTagData = (record: Tag) => {
-        setTag(record);
-        form.setFieldsValue(record);
+    const editTagData = async (record: Tag) => {
+        setLoading(true);
+
+        try {
+            const { data } = await getTagDataAPI(record.id)
+            setTag(data);
+            form.setFieldsValue(data);
+        } catch (error) {
+            setLoading(false);
+        }
+
+        setLoading(false);
     };
 
     const delTagData = async (id: number) => {
         setLoading(true);
-        await delTagDataAPI(id);
-        message.success('🎉 删除标签成功');
-        getTagList();
+
+        try {
+            await delTagDataAPI(id);
+            await getTagList();
+            message.success('🎉 删除标签成功');
+        } catch (error) {
+            setLoading(false);
+        }
     };
 
     const onSubmit = async () => {
         setLoading(true);
-        form.validateFields().then(async (values: Tag) => {
-            if (tag.id) {
-                await editTagDataAPI({ ...tag, ...values });
-                message.success('🎉 编辑标签成功');
-            } else {
-                await addTagDataAPI(values);
-                message.success('🎉 新增标签成功');
-            }
+        setBtnLoading(true);
 
-            getTagList();
-            form.resetFields();
-            form.setFieldsValue({ name: '' })
-            setTag({} as Tag);
-        });
+        try {
+            form.validateFields().then(async (values: Tag) => {
+                if (tag.id) {
+                    await editTagDataAPI({ ...tag, ...values });
+                    message.success('🎉 编辑标签成功');
+                } else {
+                    await addTagDataAPI(values);
+                    message.success('🎉 新增标签成功');
+                }
+
+                await getTagList();
+                form.resetFields();
+                form.setFieldsValue({ name: '' })
+                setTag({} as Tag);
+            });
+        } catch (error) {
+            setLoading(false);
+            setBtnLoading(false);
+        }
+
+        setBtnLoading(false);
     };
 
     return (
-        <>
+        <div>
             <Title value="标签管理" />
 
             <div className='flex md:justify-between flex-col md:flex-row mx-auto mt-2'>
@@ -81,14 +111,13 @@ const TagPage = () => {
                         initialValues={tag}
                         onFinish={onSubmit}
                         size='large'
-
                     >
                         <Form.Item label="标签名称" name="name" rules={[{ required: true, message: '标签名称不能为空' }]}>
                             <Input placeholder="请输入标签名称" />
                         </Form.Item>
 
                         <Form.Item>
-                            <Button type="primary" htmlType="submit" loading={loading} className="w-full">{tag.id ? '编辑标签' : '新增标签'}</Button>
+                            <Button type="primary" htmlType="submit" loading={btnLoading} className="w-full">{tag.id ? '编辑标签' : '新增标签'}</Button>
                         </Form.Item>
                     </Form>
                 </Card>
@@ -107,7 +136,7 @@ const TagPage = () => {
                     />
                 </Card>
             </div>
-        </>
+        </div>
     );
 };
